Test boundary conditions
Add a new test to check the correctness of the velocity fields on the boundary of a general geometry.
In doing so, several errors in the indices of setBCond were discovered (and fixed ofc)

Also, to reduce output in the case of an error, wrap the flag check into in a function and only test the total result.

+++ src/fields.h
@@ -16,7 +16,7 @@ void	fill1Dfield (REAL value, REAL *field, int size);
 void	fill2Dfield (REAL value, REAL **field, int sizeX, int sizeY);
 /* Sets every entry in field to value */
 
-void    initUVP (REAL ***U, REAL ***V, REAL ***P, int imax, int jmax, double *init);
+void    initUVP (REAL ***U, REAL ***V, REAL ***P, int imax, int jmax, REAL *init);
 /* Initialise the fields U, V and P to UI, VI and PI respectively */
 
 #endif /* FIELDS_H_ */
